changed kw in BGPChain constructor so that x86 will pass

+++ src/mopt/AlgoBGP.jl
@@ -75,7 +75,7 @@ mutable struct BGPChain <: AbstractChain
         * `min_improve`: minimally required improvement in chain `j` over chain `i` for an exchange move `j->i` to talk place.
         * `batch_size`: size of batches in which to update parameter vector.
     """
-    function BGPChain(id::Int=1,n::Int=10;m::MProb=MProb(),sig::Float64=0.5,upd::Int64=10,upd_by::Float64=0.01,smpl_iters::Int=1000,min_improve::Float64=10.0,acc_tuner::Float64=2.0,batch_size=1)
+    function BGPChain(id::Int=1,n::Int=10;m::MProb=MProb(),sig::Float64=0.5,upd::Int=10,upd_by::Float64=0.01,smpl_iters::Int=1000,min_improve::Float64=10.0,acc_tuner::Float64=2.0,batch_size=1)
         np = length(m.params_to_sample)
         this           = new()
         this.evals     = Array{Eval}(undef,n)
